chore: rename get/updateEntity queries
