+++ src/containers/entityDetail.jsx
@@ -1,6 +1,6 @@
 import { Dialog } from 'primereact/dialog'
 import { toast } from 'react-toastify'
-import { useGetEntitiesDetailsQuery } from '/src/services/getEntitiesDetails'
+import { useGetEntitiesDetailsQuery } from '../services/entity/getEntity'
 import PropTypes from 'prop-types'
 
 const EntityDetail = ({
