+++ src/pages/browser/detail/FolderDetail.jsx
@@ -4,8 +4,8 @@ import Thumbnail from '/src/containers/thumbnail'
 import AttributeTable from '/src/containers/attributeTable'
 import { getFolderTypeIcon } from '/src/utils'
 import { TagsField } from '/src/containers/fieldFormat'
-import { useUpdateEntitiesDetailsMutation } from '/src/services/updateEntitiesDetails'
-import { useGetEntitiesDetailsQuery } from '/src/services/getEntitiesDetails'
+import { useUpdateEntitiesDetailsMutation } from '../../../services/entity/updateEntity'
+import { useGetEntitiesDetailsQuery } from '../../../services/entity/getEntity'
 import StatusSelect from '/src/components/status/statusSelect'
 import usePubSub from '/src/hooks/usePubSub'
