+++ src/pages/browser/detail/TaskDetail.jsx
@@ -3,8 +3,8 @@ import AttributeTable from '/src/containers/attributeTable'
 import { getTaskTypeIcon } from '/src/utils'
 import { TagsField } from '/src/containers/fieldFormat'
 import { Panel } from '@ynput/ayon-react-components'
-import { useUpdateEntitiesDetailsMutation } from '/src/services/updateEntitiesDetails'
-import { useGetEntitiesDetailsQuery } from '/src/services/getEntitiesDetails'
+import { useUpdateEntitiesDetailsMutation } from '../../../services/entity/updateEntity'
+import { useGetEntitiesDetailsQuery } from '../../../services/entity/getEntity'
 import StatusSelect from '/src/components/status/statusSelect'
 import usePubSub from '/src/hooks/usePubSub'
