+++ src/pages/browser/detail/VersionDetail.jsx
@@ -5,8 +5,8 @@ import AttributeTable from '/src/containers/attributeTable'
 import { TagsField } from '/src/containers/fieldFormat'
 import { getFamilyIcon } from '/src/utils'
 import RepresentationList from '../RepresentationList'
-import { useUpdateEntitiesDetailsMutation } from '/src/services/updateEntitiesDetails'
-import { useGetEntitiesDetailsQuery } from '/src/services/getEntitiesDetails'
+import { useUpdateEntitiesDetailsMutation } from '../../../services/entity/updateEntity'
+import { useGetEntitiesDetailsQuery } from '../../../services/entity/getEntity'
 
 import StatusSelect from '/src/components/status/statusSelect'
 import usePubSub from '/src/hooks/usePubSub'
